@@ -1,8 +1,9 @@
-package com.example.demoweb;
+package com.example.demo2;
 
 import java.io.*;
-import javax.servlet.http.*;
-import javax.servlet.annotation.*;
+
+import jakarta.servlet.http.*;
+import jakarta.servlet.annotation.*;
 
 @WebServlet(name = "helloServlet", value = "/hello-servlet")
 public class HelloServlet extends HttpServlet {
